mizan-orchestration/requirements.txt
# Mizan Orchestration Backend Requirements
# Part of Mizan extension to ArcPay

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1

# LangGraph & AI
langgraph==0.0.34
langchain==0.1.13
langchain-anthropic==0.1.6
anthropic==0.21.3

# Database & Storage
supabase==2.4.1
psycopg2-binary==2.9.9

# HTTP & Utilities
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-core==2.16.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6

# Logging & Monitoring
structlog==24.1.0

# Type hints
typing-extensions==4.10.0
